feat: provide diversity metric based on curvature
The metric is based on curvature profiles. The idea is to compute for
every test case the curvature profile. For every meter of road the
curvature is approximated. From the profile of a test case we determine
the mean and standard deviation of the curvature profile.

The diversity of the whole selected set of test cases is computed as
follows: first, we compute the pairwise distance of the test cases mean
and std of the associating curvature profile. secondly, we compute the
mean of those distances.

refs: #6